feat: 카테고리 별 찜 상세 조회 완료

+++ backend/src/main/java/com/meong9/backend/domain/map/dto/MapLikeResponseDto.java
@@ -9,7 +9,7 @@ import java.util.List;
 @AllArgsConstructor
 @Builder
 public class MapLikeResponseDto {
-    private Integer categoryId; // 카테고리 아이디
+    private Long categoryId; // 카테고리 아이디
     private String categoryName;
     private List<MapLikePlaceDto> places;
